requirements.txt
amqp==2.2.1
asn1crypto==0.22.0
billiard==3.5.0.3
celery==4.0.2
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
cryptography==2.0.2
defusedxml==0.5.0
dj-database-url==0.4.1
Django==2.0.3
django-braces==1.11.0
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-cors-headers==2.1.0
django-filter==1.0.4
django-oauth-toolkit==1.0.0
django-rest-framework-social-oauth2==1.0.8
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
google-api-python-client==1.6.2
gunicorn==19.6.0
httplib2==0.10.3
idna==2.5
kombu==4.1.0
langid==1.1.6
Markdown==2.6.8
nltk==3.2.4
numpy==1.13.1
oauth2client==4.1.2
oauthlib==2.0.2
psycopg2==2.7.4
pyasn1==0.2.3
pyasn1-modules==0.0.9
pycparser==2.18
PyJWT==1.5.2
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.2
redis==2.10.5
requests==2.18.1
requests-oauthlib==0.8.0
rest-framework-auth0==0.4.5
rsa==3.4.2
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
uritemplate==3.0.0
urllib3==1.21.1
vine==1.1.4
whitenoise==3.2
